Update "Is it an online program?"
Old version:
Digital Product School (DPS) is all about teamwork. You learn how to work together in an interdisciplinary team by creating ideas, structuring your work, and making decisions. We believe that’s only possible when you are physically together in a team space that is why we don’t offer it as an online program.